requirements.txt
streamlit>=1.28.0
pandas>=2.1.0
spotipy>=2.23.0
requests>=2.31.0
python-dotenv>=1.0.0
plotly>=5.18.0